add member to task stimulus structure completed
