--- app/javascript/controllers/task_member_controller.js
@@ -1,46 +1,0 @@
-import { Controller } from "stimulus"
-import { csrfToken } from "@rails/ujs";
-
-export default class extends Controller {
-  static targets = [ "dropdown" ]
-
-  connect() {
-    console.log("task member controller connected");
-  }
-  //  dropdown for the add members function
-  revealContent() {
-    
-    const url = "/friends"
-    fetch(url, { headers: { 'Accept': 'text/plain' } })
-      .then(response => response.text())
-      .then(this.updateAndHide.bind(this))
-  }
-
-  updateAndHide(members) {
-    this.dropdownTarget.classList.toggle("hidden");
-    this.dropdownTarget.innerHTML = ""
-    this.dropdownTarget.insertAdjacentHTML('afterbegin', members )
-  }
-
-
-  // function transferred to task_controller.js for bugfixing
-
-  // add(e) {
-  //   // console.log(this.memberTarget);
-  //   e.preventDefault();
-  //   const url = `/tasks/${ this.element.dataset.taskId }/task_members/`
-  //   // how to pass in the params of user??
-  //   fetch(url, { 
-  //       method: 'POST', 
-  //       headers: { 'Accept': 'text/plain', 'X-CSRF-token': csrfToken() }, 
-  //       body: JSON.stringify({ task_member_id: this.membernamesTarget.dataset.memberId })
-  //   })
-  //   .then(response => response.text())
-  //   .then(this.addMember.bind(this))
-  // }
-
-  // addMember(member) {
-  //   console.log(member)
-  //   this.memberTarget.insertAdjacentHTML( 'afterbegin', member.first_name )
-  // }
-}
