+++ app/views/members/_add_member.html.erb
@@ -1,20 +1,17 @@
 <%# this button is to be added to the task partial %>
 
 <%# THIS COMPONENT IS TAKING A HARD CODED TASK ID =>  data-task-id="1" %>
-<div class="w-48" data-controller="task-member">
-  <div data-task-target="member" class="shadow-lg p-2 rounded-full inline-block">
-    <%# member first_name displays here %>
-  </div>
+<div data-controller="toggle">
 
   <%# button to add member %>
-  <div data-action="click->task-member#revealContent" class="shadow-lg p-2 rounded-full inline-block">
+  <div data-action="click->toggle#dropdownAddMembers" class="shadow-lg p-2 rounded-full inline-block cursor-pointer">
     <svg xmlns="http://www.w3.org/2000/svg" width="15.998" height="16" viewBox="0 0 15.998 16">
       <path id="profile_close_add" d="M363.994,367v-1h-1a1,1,0,1,1,0-2h1v-1a1,1,0,0,1,2,0v1h1a1,1,0,1,1,0,2h-1v1a1,1,0,0,1-2,0Zm-10.972-.97a2.528,2.528,0,0,1,0-4.06,9.654,9.654,0,0,1,2.512-1.35l.01-.009a5.15,5.15,0,0,1-.46-.52c-.024-.031-.056-.058-.08-.091h0a4.951,4.951,0,0,1-1.026-3,5.006,5.006,0,1,1,8.45,3.611l.009.009a5.146,5.146,0,0,1-.431.339c-.088.067-.168.144-.262.205a5.122,5.122,0,0,1-.5.272c-.115.059-.223.129-.343.178a4.953,4.953,0,0,1-.6.186c-.11.03-.213.077-.328.1a5.01,5.01,0,0,1-.993.1,7.945,7.945,0,0,0-4.766,1.58.514.514,0,0,0,0,.84,8.1,8.1,0,0,0,4.755,1.59.991.991,0,0,1,1.012.99,1.01,1.01,0,0,1-1.031,1A9.925,9.925,0,0,1,353.022,366.03Zm2.963-9.03a3,3,0,1,0,3-3A3,3,0,0,0,355.985,357Z" transform="translate(-352.001 -352)"/>
     </svg>
   </div>
 
     <%# Dropdown menu, show/hide based on menu state. %>
-  <form data-task-member-target="dropdown" class="shadow-lg rounded-lg hidden" role="menu">
+  <form data-task-Id="<%= task.id %>" data-task-target="taskId" data-toggle-target="dropdown" class="fixed bg-white w-48 shadow-lg rounded-lg hidden">
     <%#  serach bar for users goes here %>
     <%# trying to work with 1 user as a submit button at first %>
 </div>
