--- app/views/members/_members.html.erb
@@ -1,7 +1,0 @@
-
-<% members.each do |member| %>
-  <button data-task-target="membernames" data-member-id=<%= member.id %> type="submit" data-action="click->task#add" class="w-full mt-1 px-4 py-2 text-sm hover:bg-gray-100 duration-100">
-    <%= member.first_name.capitalize %>&nbsp;<%= member.last_name.capitalize %>
-  </button>
-<% end %>
-
