+++ app/views/members/_members_to_add.html.erb
@@ -1,0 +1,7 @@
+
+<% members.each do |member| %>
+  <button data-task-target="memberAddDropdownName" data-member-id=<%= member.id %> data-action="click->task#addMember" class="w-full mt-1 px-4 py-2 text-sm hover:bg-gray-100 duration-100">
+    <%= member.first_name.capitalize %>&nbsp;<%= member.last_name.capitalize %>
+  </button>
+<% end %>
+
