+++ app/views/shared/_sidebar_l.html.erb
@@ -9,7 +9,7 @@
   </button>
 
   <%# create button dropdown %>
-  <div data-sidebar-l-target="dropdown" class="t-dropdown w-48 hidden">
+  <div data-sidebar-l-target="dropdownCreate" class="t-dropdown w-48 hidden">
     <%= link_to 'Workflow', workflows_path, method: 'post',
     class: "t-dropdown-btn" %>
     <button data-action="click->world#createWorkflow" class="block w-full flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">
