+++ app/views/tasks/_task.html.erb
@@ -1,17 +1,17 @@
-<div data-controller="toggle-task">
+<div data-controller="toggle">
   <div data-task-id=<%= task.id %> data-workflow-id=<%= workflow.id %> data-controller="item" class="t-card border py-4 px-4 mb-2 mx-6">
     <div class="flex justify-between">
-      <button class="text-sm flex-grow text-left" data-action="click->item#getItems click->toggle-task#hide">
+      <button class="text-sm flex-grow text-left" data-action="click->item#getItems click->toggle#hide">
         <%= task.title %>
       </button>
 
-      <%= render partial: 'members/add_member' %>
+      <%= render partial: 'members/add_member', locals: {task: task} %>
 
       <button data-action="click-> task#markComplete" class='inline-block text-indigo-500 text-sm'>
         Mark as complete
       </button>
     </div>
-    <div class="hidden" data-toggle-task-target="hide" >
+    <div class="hidden" data-toggle-target="hide" >
 
       <ul data-item-target="list" >
       </ul>
